Create vlad buiding beautiful array
Vlad was given an array a of n positive integers. Now he wants to build a beautiful array b
 of length n from it.

Vlad considers an array beautiful if all the numbers in it are positive and have the same parity. That is, all numbers in the beautiful array are greater than zero and are either all even or all odd.
To build the array b, Vlad can assign each bi either the value ai
or ai−aj, where any j from 1 to n can be chosen.
To avoid trying to do the impossible, Vlad asks you to determine whether it is possible to build a beautiful array b of length n using his array a